More markbar functions, fixed markbar width/height
Attempt to write test cases for constant markbar size even when
making/closing splits, but those seem to fail(?) even though that
functionality works properly when I test it myself in a vim instance.

Add convenience functions for closing the current markbar, getting the
markbar window number (if it's open). Rename a function to make it more
obvious that it's an 'getter' (rather than a 'setter').