Callee-save ignorecase in 'Is[...]Mark()' funcs
vim-markbar uses `match()` to determine whether a mark (e.g. `'a`) is a
"local" mark or a file mark. When `ignorecase` is set, `'a` (and other
lowercase letters) will (incorrectly) match against their uppercase
equivalents in the `helpers#IsGlobalMark()` and
`helpers#IsUppercaseMark()` helper functions.

This issue caused vim-markbar to throw 'Key not found' exceptions when
users tried to open the markbar.

@ref Issue 1 (https://github.com/Yilin-Yang/vim-markbar/issues/1)