Make syntax work w/ zero-height section separator
Previously, markbar syntax would match whitespace-indented text into a
'markbarContext' syntax *region* whose "end" pattern was a blank line.
This felt more elegant than individually `syn match`ing each indented
line, and it  made it possible to fold mark "sections" in the markbar.

The reason I didn't use the start of the next "mark section heading" as
the `end` pattern for the `markbarContext` region was that it would
interfere with the matching for the `markbarSectionHeader` region. As it
turns out, you can "scooch" the match to the left by specifying an
offset, which lets me terminate the `markbarContext` region without
"eating" part of the `markbarSectionHeader` region.