Set `nobuflisted` after first entry into markbar
It seems(?) that vim sets `buflisted` to true the first time the buffer
is entered, which clobbers the `nobuflisted` value that gets set in
`ui#SetMarkbarBufferSettings`. Setting this option after the markbar
split has been open seems to fix things.

Add test case to check that markbar buffer settings are set properly.