Create markbar buffer with bufnr, not badd
Tentative fix for #3. Running hypothesis is that `badd` is successfully
creating a new buffer for the markbar, but that the following call to
`bufnr` can't find it in the buffer list (thereby returning '-1').

Try to circumvent this by creating the markbar buffer using `bufnr`, and
using its return value. Also, default-initialize
MarkbarView::_markbar_buffer to -2 instead of -1, to differentiate
between `badd` failure and `bufnr` failure.