Add FetchBufferLineRange's sed cmd in a try-catch
Trying to read lines from a file using a `system('sed...` call can fail
noisily (E484) if the given file doesn't exist (e.g. if the given buffer
corresponds to a stale tempfile). Wrap that call in a try-catch, so that
execution can proceed and the function can "try something else."