"Redundantly" create jump-to-mark mappings
vim-markbar maps single-quote and backtick to "open peekaboo markbar,"
which can produce a noticeable and unnecessary delay when jumping
directly to a "known" mark (e.g. double-tapping single-quote to go to
last jump position).

Add an option to explicitly create mappings comparable to:

    noremap 'a  :normal! 'a<cr>
    noremap 'b  :normal! 'b<cr>
    ...
    noremap `z  :normal! `z<cr>

So that ' and ` only open the peekaboo markbar when the user
"hesitates."

Addresses Issue #6.