Work around occasional 'Key not present in dict'
Markbar would throw occasional 'E716: Key not present in dict'
exceptions (due to a buffer not having had its marks cached) when trying
to open the markbar immediately after opening vim.

The issue seems to be that, on VimEnter, markbar erroneously registers
an unlisted, never-actually-been-opened buffer (usually bufno 2 or 3) as
the "most recently opened buffer." Because this buffer hasn't been
opened, markbar hasn't cached its marks, and the attempted cache lookup
throws a 'Key not present in dict'.

Work around this by detecting when this happens, closing the nascent
markbar, forcibly registering the buffer currently open as the 'active
buffer', and then reopening the markbar.