Open markbar buffer with `buffer!`, not `split #N`
Problem:

The exception thrown in Issue #12 occurs from within
`MarkbarView::_openMarkbarSplit`, which opens a buffer *different* from
the markbar buffer after following the steps to reproduce.

The `[:edit|:split] #N` syntax tries to open buffer number `N` as if it
were a file. After setting a local working directory (with `:lcd` or
`:tcd`), `:split #N` will treat the markbar buffer's name
(which defaults to `( Markbar)` ) as an actual filename, given relative
to the new local working directory.

For instance, if the LWD is `~/plugin/vim-markbar` and N = 2, then
`:split #2` will try to open a buffer with the filename
`~/plugin/vim-markbar/( Markbar )`; since this doesn't match
`( Markbar )`, this creates a *new* buffer instead of opening buffer 2
as intended. This may be an upstream bug with vim itself.

Solution:

Open the markbar split, in `MarkbarView::_openMarkbarSplit`, using
`split` followed by `buffer!`, which apparently doesn't have this
problem.