Gracefully handle "mark not found" with [''] mark
This issue only seems to occur when markbar can't retrieve the last
active buffer, and falls back to only showing file marks to the user:
when this is true, if the user tries to go to mark `'`, markbar would
throw an exception and an ugly, multiline stacktrace instead of showing
a 'Mark not found' error.

Ideally, I should try to figure out why markbar isn't pushing buffers
onto the stack in these cases, but in the meantime, make these errors
less obtrusive.