Properly retrieve and display 'special' marks
vim-markbar wasn't actually capable of caching and displaying certain
marks, like: '^, '), '[, and so on. Refactor `helpers#GetLocalMarks()`
(and write a helper function) so that it does.

Fix syntax highlighting for these special marks.

Fix the use of function references to generate default mark names:
this hadn't been tested with local marks, and when used with local
marks, `helpers#ParentFilename()` would fail to retrieve the correct
buffer name. Tweak the corresponding test case in
`test-StandardMarkbarController-names` to check for this.

Add some miscellaneous style fixes.

Addresses Issue #13.