Restore old window layout on markbar close
When, e.g. the same buffer is open in multiple windows, and the user
opens (but then subsequently closes) the markbar, the cursor will not
return to the previously active window. (Also, the sizes of the windows
open may change when the markbar closes.) Save the window layout and
the last active window, then restore it on closing the markbar.

Also, in the scenario above, move the cursor to the last active window
when actually jumping to a mark.