Fix marks being opened inside the markbar itself
Previously, the markbar was "refreshed" by calling openMarkbar() from
inside the markbar.

This had the (desired) effect of forcibly clearing and repopulating the
markbar's contents.

This had the undesired effect of clobbering the "last window position"
that was set when the markbar was first opened, overwriting it with the
window number of the markbar itself. When the user tried jumping to a
mark, the target buffer would open inside the markbar.

Solution:
- Write a separate refreshContents() function, and invoke that to update
the markbar.
- Don't save the current window state if the active buffer is a markbar.

Also,

Changes to openMarkbar() break a few of the StandardMarkbarController
tests; closing the markbar, in those tests, no longer returns to the
active buffer, but instead to the Vader-workbench.

This problem doesn't occur in the functional tests; the working
assumption is that this is a quirk of vader, and not a true defect.