Allow diff num_lines_context for diff mark types
Allow users to request different numbers of lines of context for
local vs. file marks, in the standard and peekaboo markbars.

- Change context retrieval logic to grab the maximum number of lines
requested by the user. Determine which lines to render based on
mark-/markbar-type when the markbar is shown.
- Allow FetchContext to return zero lines of context (so that users can
show no context for, e.g. file marks).
- Write and pass tests for the above.
- Separate "open markbar" test cases into standalone files, since they
would each configure settings that affected the outcome of later tests.
- Document changes to g:markbar_num_lines_context in docs and README.
- Miscellaneous style fixes.

Closes #14.