Open folds like foldopen when jumping to a mark
Jumping to a mark from either markbar did not open the fold(s) in which
the mark was located (see issue #18).

Add a `g:markbar_foldopen` setting that, when true, causes vim-markbar
to open the mark's fold(s) if it's located in one. Add corresponding
tests and documentation.

Also: fix a silly style error in `load-test-files.vader`.