Activate last-active window with win_gotoid()
Before jumping to a user's mark, the previous implementation would
store the winnr() of the last-active window and return to that window
with `wincmd w`. This fails when g:markbar_open_position is 'topleft'
because the newly opened markbar would have winnr() 1, and the winnr of
all other windows would be incremented.

Instead store the win_getid() of the last-active window and return to it
with win_gotoid().

Fixes #27.