Optimize up markbar#helpers#FetchBufferLineRange()
Repeated calls to readfile() can be very slow when the marked files are
very large. Cache file contents between FetchBufferLineRange() calls by
loading the file in a hidden buffer. Only use readfile() when the user
unloaded or deleted the buffer.

"Caching" a file will still be slow if the file is very, very large, but
it seems necessary to pay that performance penalty if a user is editing
and setting marks in a file that large.